+++ app/src/main/kotlin/me/xizzhu/android/joshua/search/result/SearchVerseItem.kt
@@ -17,11 +17,16 @@
 package me.xizzhu.android.joshua.search.result
 
 import android.annotation.SuppressLint
+import android.graphics.Color
 import android.text.SpannableStringBuilder
+import android.text.style.BackgroundColorSpan
+import android.text.style.ForegroundColorSpan
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.TextView
+import androidx.annotation.ColorInt
 import me.xizzhu.android.joshua.R
+import me.xizzhu.android.joshua.core.Highlight
 import me.xizzhu.android.joshua.core.Settings
 import me.xizzhu.android.joshua.core.VerseIndex
 import me.xizzhu.android.joshua.ui.*
@@ -31,6 +36,7 @@ import java.util.*
 
 data class SearchVerseItem(val verseIndex: VerseIndex, private val bookShortName: String,
                            private val text: String, private val query: String,
+                           @ColorInt private val highlightColor: Int,
                            val onClicked: (VerseIndex) -> Unit)
     : BaseItem(R.layout.item_search_verse, { inflater, parent -> SearchVerseItemViewHolder(inflater, parent) }) {
     companion object {
@@ -68,6 +74,15 @@ data class SearchVerseItem(val verseIndex: VerseIndex, private val bookShortName
                         if (index == 0) KEYWORD_STYLE_SPAN else createKeywordStyleSpan(),
                         start, start + keyword.length)
             }
+        }
+
+        // highlights the verse if needed
+        if (highlightColor != Highlight.COLOR_NONE) {
+            SPANNABLE_STRING_BUILDER.setSpan(
+                    BackgroundColorSpan(highlightColor),
+                    ForegroundColorSpan(if (highlightColor == Highlight.COLOR_BLUE) Color.WHITE else Color.BLACK),
+                    SPANNABLE_STRING_BUILDER.length - text.length, SPANNABLE_STRING_BUILDER.length
+            )
         }
 
         return@lazy SPANNABLE_STRING_BUILDER.toCharSequence()